Update `Timeout` to `double`

+++ ricaun.RevitTest.Command/Process/RevitTestProcessStart.cs
@@ -48,7 +48,7 @@ namespace ricaun.RevitTest.Command.Process
             if (!close) return this;
             return SetRevitArgument("close");
         }
-        public RevitTestProcessStart SetTimeout(int timeoutMinutes) => SetRevitArgument("timeout", timeoutMinutes);
+        public RevitTestProcessStart SetTimeout(double timeoutMinutes) => SetRevitArgument("timeout", timeoutMinutes);
         public RevitTestProcessStart SetTestFilter(string[] testFilters)
         {
             if (testFilters.Length == 0)
